+++ ui/animation/animate.js
@@ -1,0 +1,8 @@
+jQuery.fn.extend({
+    animateCss: function (animationName) {
+        var animationEnd = 'webkitAnimationEnd mozAnimationEnd MSAnimationEnd oanimationend animationend';
+        this.addClass('animated ' + animationName).one(animationEnd, function() {
+            jQuery(this).removeClass('animated ' + animationName);
+        });
+    }
+});